+++ vk-driver/src/dev/endpoints.rs
@@ -484,7 +484,7 @@ mod tests {
             exec_env: vec![],
             endpoints,
             host_exec: None,
-            ssh_agent: false,
+            ssh: Default::default(),
             cache: Default::default(),
             requires: Default::default(),
             cached_only: false,