+++ vk-driver/src/dev/mod.rs
@@ -28,6 +28,7 @@ pub mod list;
 pub mod plan;
 pub mod schema;
 mod session;
+pub mod sshsetup;
 mod status;
 pub mod storage;
 mod table;
@@ -147,7 +148,7 @@ pub(super) mod testutil {
             exec_env: vec![],
             endpoints: vec![],
             host_exec: None,
-            ssh_agent: false,
+            ssh: Default::default(),
             cache: Default::default(),
             requires: Default::default(),
             cached_only: false,